Move all network handling code to the correct cog
